Feature : [#1] separate service layer
